sd3/requirements.txt
# SD3 open_source (davae + omini) 依赖合集（按需安装）

# core
torch
torchvision
accelerate
omegaconf
diffusers
transformers
huggingface_hub

# training
lightning
peft
tensorboard

# data / yaml
webdataset
iopath
ruamel.yaml
pyyaml

# image / math
Pillow
numpy
matplotlib
tqdm
scikit-image
albumentations
kornia

# tokenizer/metrics (按你仓库实际使用情况保留)
open_clip_torch
elatentlpips

# 若启用 datasets（某些旧脚本会 import），需要这些常见依赖
datasets
pandas
python-dateutil


